♻️  Improved credential handling
